+++ DataLayers/Framework/ModIntegrations.cs
@@ -14,6 +14,13 @@ namespace Pathoschild.Stardew.DataLayers.Framework
     internal class ModIntegrations
     {
         /*********
+        ** Fields
+        *********/
+        /// <summary>An API for fetching metadata about loaded mods.</summary>
+        private readonly IModRegistry ModRegistry;
+
+
+        /*********
         ** Accessors
         *********/
         /// <summary>Handles access to the Automate mod.</summary>
@@ -50,6 +57,8 @@ namespace Pathoschild.Stardew.DataLayers.Framework
         /// <param name="reflection">An API for accessing private code.</param>
         public ModIntegrations(IMonitor monitor, IModRegistry modRegistry, IReflectionHelper reflection)
         {
+            this.ModRegistry = modRegistry;
+
             this.Automate = new AutomateIntegration(modRegistry, monitor);
             this.BetterJunimos = new BetterJunimosIntegration(modRegistry, monitor);
             this.BetterSprinklers = new BetterSprinklersIntegration(modRegistry, monitor);
@@ -58,6 +67,13 @@ namespace Pathoschild.Stardew.DataLayers.Framework
             this.PelicanFiber = new PelicanFiberIntegration(modRegistry, reflection, monitor);
             this.PrismaticTools = new PrismaticToolsIntegration(modRegistry, monitor);
             this.SimpleSprinkler = new SimpleSprinklerIntegration(modRegistry, monitor);
+        }
+
+        /// <summary>Get whether a mod has been loaded.</summary>
+        /// <param name="uniqueID">The mod's unique ID.</param>
+        public bool HasMod(string uniqueID)
+        {
+            return this.ModRegistry.IsLoaded(uniqueID);
         }
     }
 }