+++ DataLayers/Layers/Coverage/BeeHouseLayer.cs
@@ -25,7 +25,7 @@ namespace Pathoschild.Stardew.DataLayers.Layers.Coverage
         private readonly int MaxRadius = 5;
 
         /// <summary>The relative tile coordinates covered by a bee house.</summary>
-        private readonly Vector2[] RelativeRange = BeeHouseLayer.GetRelativeCoverage().ToArray();
+        private readonly Vector2[] RelativeRange;
 
 
         /*********
@@ -34,13 +34,18 @@ namespace Pathoschild.Stardew.DataLayers.Layers.Coverage
         /// <summary>Construct an instance.</summary>
         /// <param name="translations">Provides translations in stored in the mod folder's i18n folder.</param>
         /// <param name="config">The data layer settings.</param>
-        public BeeHouseLayer(ITranslationHelper translations, LayerConfig config)
+        /// <param name="hasBeeHouseFlowerRangeFix">Whether the player has the Bee House Flower Range Fix mod.</param>
+        public BeeHouseLayer(ITranslationHelper translations, LayerConfig config, bool hasBeeHouseFlowerRangeFix)
             : base(translations.Get("bee-houses.name"), config)
         {
             this.Legend = new[]
             {
                 new LegendEntry(translations.Get("bee-houses.range"), this.CoveredColor)
             };
+
+            this.RelativeRange = BeeHouseLayer
+                .GetRelativeCoverage(maxLoops: hasBeeHouseFlowerRangeFix ? 166 : 150)
+                .ToArray();
         }
 
         /// <summary>Get the updated data layer tiles.</summary>
@@ -105,13 +110,14 @@ namespace Pathoschild.Stardew.DataLayers.Layers.Coverage
         }
 
         /// <summary>Get the relative tiles covered by a bee house.</summary>
+        /// <param name="maxLoops">The maximum number of flood fill loops to run, which affects the range. The game uses 150 by default.</param>
         /// <remarks>Derived from <see cref="Utility.findCloseFlower"/>.</remarks>
-        private static IEnumerable<Vector2> GetRelativeCoverage()
+        private static IEnumerable<Vector2> GetRelativeCoverage(int maxLoops = 150)
         {
             Queue<Vector2> queue = new Queue<Vector2>();
             HashSet<Vector2> visited = new HashSet<Vector2>();
             queue.Enqueue(Vector2.Zero);
-            for (int i = 0; i <= 150 && queue.Count > 0; ++i)
+            for (int i = 0; i <= maxLoops && queue.Count > 0; ++i)
             {
                 Vector2 tile = queue.Dequeue();
                 yield return tile;